+++ Tests/qualite.py
@@ -1,6 +1,6 @@
 """actions qualite
 
-- badge de comptage de fonctions
+- badge de comptage de fonctions et lignes de code
 """
 from pathlib import Path
 import anybadge
@@ -13,7 +13,7 @@ sys.path.insert(0, str(myDir))  # translate from posixPath
 from Code import marss
 from inspect import getmembers, isfunction
 
-nombre_fonctions = len(getmembers(marss,isfunction))
+nombre_fonctions = len(getmembers(marss, isfunction))
 badge = anybadge.Badge('fonctions', nombre_fonctions, default_color='green')
 badge.write_badge('Doc/fonctions.svg', overwrite=True)